Eval: Do not automatically create a new env when evaluating a block
Previously evaluating a block evaluated all of its statements in a
new child env. This is good for almost everything, except the 'for'
statement where the pre, condition and post statements need to run
in the same env. Therefore now we need to create a new child env
and pass that one when evaluating a block. This allows us to fix
for statements finally